Seperate all styles components (#9)
* refactor style cart-dropdown

* refactor styles cart-icon

* refactor styles cart-item

* refactor styles category-container

* refactor styles category-item

* refactor styles category-preview

* refactor styles checkout-item

* refactor styles form-input

* refactor styles and seperate utils sign-in

* refactor styles and seperate utils sign-up

* refactor styles navigation

* refactor styles payment-form

* refactor styles product-card
