+++ src/components/category/category-container/styles.jsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const CategoryContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  justify-content: space-between;
+`;
