+++ src/components/checkout-item/index.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import styled, { css } from "styled-components";
+
 import { useDispatch, useSelector } from "react-redux";
 
 import { selectCartItems } from "store/cart/cart-selector";
@@ -9,58 +9,17 @@ import {
   clearItemFromCart,
 } from "store/cart/cart-action";
 
-const CheckoutItemContainer = styled.div`
-  width: 100%;
-  display: flex;
-  min-height: 100px;
-  border-bottom: 1px solid darkgrey;
-  padding: 15px 0;
-  font-size: 20px;
-  align-items: center;
-`;
-
-const ImageContainer = styled.div`
-  width: 23%;
-  padding-right: 15px;
-`;
-
-const Image = styled.img.attrs((props) => ({
-  src: props.src,
-  alt: `item ${props.name}`,
-}))`
-  width: 100%;
-  height: 100%;
-`;
-
-const textStyle = css`
-  width: 23%;
-`;
-
-const TextName = styled.span`
-  ${textStyle}
-`;
-
-const QuantityContainer = styled.span`
-  ${textStyle}
-  display: flex;
-`;
-
-const Arrow = styled.div`
-  cursor: pointer;
-`;
-
-const TextQuantity = styled.span`
-  margin: 0 10px;
-`;
-
-const TextPrice = styled.span`
-  ${textStyle}
-`;
-
-const RemoveButton = styled.div`
-  padding-left: 12px;
-  cursor: pointer;
-`;
+import {
+  CheckoutItemContainer,
+  Arrow,
+  ImageContainer,
+  Image,
+  QuantityContainer,
+  TextName,
+  TextQuantity,
+  TextPrice,
+  RemoveButton,
+} from "./styles";
 
 const CheckoutItem = ({ cartItem }) => {
   const dispatch = useDispatch();
