+++ src/components/form/sign-in/utils.jsx
@@ -1,0 +1,4 @@
+export const defaultFormFields = {
+  email: "",
+  password: "",
+};
