+++ src/components/form/sign-up/styles.jsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const colorGray = "gray";
+
+export const SignUpContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 480px;
+`;
+
+export const Title = styled.h2`
+  margin: 10px 0;
+`;
+
+export const Description = styled.span`
+  color: ${colorGray};
+`;
