+++ src/components/form/sign-up/utils.jsx
@@ -1,0 +1,6 @@
+export const defaultFormFields = {
+  displayName: "",
+  email: "",
+  password: "",
+  confirmPassword: "",
+};
