+++ src/components/navigation/index.jsx
@@ -1,8 +1,6 @@
 import { useSelector, useDispatch } from "react-redux";
-
 import { AnimatePresence } from "framer-motion";
-import styled from "styled-components";
-import { Outlet, NavLink } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 
 import { signOutStart } from "store/user/user.action";
 import { selectCurrentUser } from "store/user/user.selector";
@@ -10,45 +8,13 @@ import { selectIsCartOpen } from "store/cart/cart-selector";
 
 import { ReactComponent as CrwnLogo } from "assets/logo/crown.svg";
 
+import {
+  NavigationContainer,
+  LogoContainer,
+  OptionsContainer,
+  OptionLink,
+} from "./styles";
 import { CartIcon, CartDropdown } from "components/cart";
-
-const NavigationContainer = styled.div`
-  height: 70px;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  padding: 0 80px;
-`;
-
-const LogoContainer = styled(NavLink)`
-  height: 100%;
-  width: 80px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const OptionsContainer = styled.div`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-`;
-
-const OptionLink = styled(NavLink)`
-  padding: 10px 15px;
-  cursor: pointer;
-  font-weight: 500;
-  text-transform: uppercase;
-  &.active {
-    color: #696969;
-  }
-
-  &:hover {
-    text-decoration: underline;
-  }
-`;
 
 const Navigation = () => {
   const dispatch = useDispatch();
