+++ src/components/product-card/styles.jsx
@@ -1,0 +1,58 @@
+import styled from "styled-components";
+import { motion } from "framer-motion";
+
+import Button from "components/button";
+
+export const ProductImage = styled.img`
+  width: 100%;
+  height: 95%;
+  object-fit: cover;
+  margin-bottom: 5px;
+`;
+
+export const ProductCardContainer = styled(motion.div)`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  height: 350px;
+  align-items: center;
+  position: relative;
+
+  button {
+    width: 80%;
+    opacity: 0.7;
+    position: absolute;
+    top: 255px;
+    display: none;
+  }
+
+  &:hover {
+    ${ProductImage} {
+      opacity: 0.8;
+    }
+
+    button {
+      opacity: 0.85;
+      display: flex;
+    }
+  }
+`;
+
+export const ProductCardFooter = styled.div`
+  width: 100%;
+  height: 5%;
+  display: flex;
+  justify-content: space-between;
+  font-size: 18px;
+`;
+
+export const ProductName = styled.span`
+  width: 90%;
+  margin-bottom: 15px;
+`;
+
+export const ProductPrice = styled.span`
+  width: 10%;
+`;
+
+export const ProductButton = styled(Button);
